Organizes app directory
