+++ Android/Android Apps/MemeGenerator/app/src/main/java/co/ivanebernal/memegenerator/model/TextProperties.java
@@ -1,4 +1,4 @@
-package co.ivanebernal.memegenerator;
+package co.ivanebernal.memegenerator.model;
 
 public class TextProperties {
     private String text;
