+++ Android/Android Apps/MemeGenerator/app/src/main/java/co/ivanebernal/memegenerator/util/SimpleTextWatcher.java
@@ -1,4 +1,4 @@
-package co.ivanebernal.memegenerator;
+package co.ivanebernal.memegenerator.util;
 
 import android.text.Editable;
 import android.text.TextWatcher;
